Pascal's triangle printing
Pascal's triangle is a triangular array of the binomial coefficients. It is constructed by adding the numbers to the left and right of the number above to find the new value to place in that position. If either the number to the right or left is not present, substitute a zero in its place.